Dat be

+++ backend/user-service/src/main/java/com/userservice/service/UserService.java
@@ -7,11 +7,11 @@ import org.springframework.http.ResponseEntity;
 public interface UserService {
     ResponseEntity<?> findByUsername(String username);
     ResponseEntity<?> save(FormReg formReg);
-    ResponseEntity<?> findById(int id);
+    ResponseEntity<?> findById(long id);
     ResponseEntity<?> getAll(int page,String sort);
     Boolean existsByUsername(String username);
     Boolean existsByEmail(String email);
     User findByEmail(String email);
 
-    ResponseEntity<?> activeToLecture(int id);
+    ResponseEntity<?> activeToLecture(long id);
 }
